non funzione il sndMsgBT

+++ green_house_controller/greenhousecontrollerino/TaskSearch.cpp
@@ -12,7 +12,7 @@ void TaskSearch::init(int period){
 
 void TaskSearch::tick(){
   Serial.print("searching....");
-  int a = this->prox->getValue();
+  float a = this->prox->getValue();
   Serial.println(a);
   if(!GLOBAL_CLASS.isConnected() && a < GLOBAL_CLASS.getEngagementDist()){
     GLOBAL_CLASS.toggleConnection();
